requirements.txt
langchain==0.3.27
openai==1.109.1
chromadb==1.1.0
sentence-transformers>=2.2.0
pypdf==6.1.0
python-dotenv>=1.0.0
tqdm>=4.64.0
